Move significance logic to shared package
